bedrock-api/requirements.txt
fastapi
uvicorn[standard]
boto3
pydantic
python-dotenv

